mobile optimizations
