+++ src/app/config/badge-components.config.ts
@@ -1,10 +1,10 @@
 import { ComponentStructure } from './app.config';
-import { BadgeContentTypeComponent } from './components/ui/badge/badge-content-type/badge-content-type.component';
-import { BadgeCustomComponent } from './components/ui/badge/badge-custom/badge-custom.component';
-import { BadgeDiComponent } from './components/ui/badge/badge-di/badge-di.component';
-import { BadgeLongComponent } from './components/ui/badge/badge-long/badge-long.component';
-import { BadgeSizesComponent } from './components/ui/badge/badge-sizes/badge-sizes.component';
-import { BadgeComponent } from './components/ui/badge/badge/badge.component';
+import { BadgeContentTypeComponent } from '../components/ui/badge/badge-content-type/badge-content-type.component';
+import { BadgeCustomComponent } from '../components/ui/badge/badge-custom/badge-custom.component';
+import { BadgeDiComponent } from '../components/ui/badge/badge-di/badge-di.component';
+import { BadgeLongComponent } from '../components/ui/badge/badge-long/badge-long.component';
+import { BadgeSizesComponent } from '../components/ui/badge/badge-sizes/badge-sizes.component';
+import { BadgeComponent } from '../components/ui/badge/badge/badge.component';
 
 export const badgeComponents: Omit<ComponentStructure, 'id'>[] = [
   {
@@ -15,35 +15,35 @@ export const badgeComponents: Omit<ComponentStructure, 'id'>[] = [
     component: BadgeComponent,
   },
   {
-    label: 'Badges',
-    title: 'Badge sizes',
+    label: 'Badges Sizes',
+    title: 'Badge Sizes',
     description: 'User can set sizes',
     url: 'https://taiga-ui.dev/components/badge#sizes',
     component: BadgeSizesComponent,
   },
   {
-    label: 'Badges',
-    title: 'Badge content type',
+    label: 'Badges Content Type',
+    title: 'Badge Content Type',
     description: 'User can provide icons and images',
     url: 'https://taiga-ui.dev/components/badge#content-type',
     component: BadgeContentTypeComponent,
   },
   {
-    label: 'Badges',
-    title: 'Badge long',
+    label: 'Badges Long',
+    title: 'Badge Long',
     description: 'User can project long text',
     url: 'https://taiga-ui.dev/components/badge#long-value',
     component: BadgeLongComponent,
   },
   {
-    label: 'Badges',
-    title: 'Badge customization',
+    label: 'Badges Customization',
+    title: 'Badge Customization',
     description: 'User can customize appearance',
     url: 'https://taiga-ui.dev/components/badge#customization',
     component: BadgeCustomComponent,
   },
   {
-    label: 'Badges',
+    label: 'Badges With DI',
     title: 'Badge With DI',
     description: 'Badge configured with Dependency Injection',
     url: 'https://taiga-ui.dev/components/badge#options',
