+++ src/app/dashboard/dashboard.component.ts
@@ -5,6 +5,7 @@ import { DashboardService } from "../services/dashboard.service";
 import { NgComponentOutlet } from "@angular/common";
 import { ElementContainerComponent } from "../controls/element-container/element-container.component";
 import { ElementCardComponent } from "../controls/element-card/element-card.component";
+import {TuiHeader, TuiNavigation} from '@taiga-ui/layout';
 
 @Component({
     selector: 'app-dashboard',
@@ -12,7 +13,7 @@ import { ElementCardComponent } from "../controls/element-card/element-card.comp
     templateUrl: './dashboard.component.html',
     styleUrl: 'dashboard.component.scss',
     changeDetection: ChangeDetectionStrategy.OnPush,
-    imports: [ToolbarComponent, ButtonComponent, NgComponentOutlet, ElementContainerComponent, ElementCardComponent],
+    imports: [TuiHeader, TuiNavigation, ToolbarComponent, ButtonComponent, NgComponentOutlet, ElementContainerComponent, ElementCardComponent],
 })
 export class DashboardComponent {
