+++ src/styles/_utils.scss
@@ -10,6 +10,10 @@
   padding: 16px;
 }
 
+.w-full {
+  width: 100%;
+}
+
 /*tmp workaround*/
 .tui-space_left-1 {
   margin-left: 0.5rem;
